+++ server/paymentProviders/stripe/creditcard.ts
@@ -1,16 +1,9 @@
 import config from 'config';
 import { get, toUpper } from 'lodash';
 import type Stripe from 'stripe';
-import { v4 as uuid } from 'uuid';
 
-import FEATURE from '../../constants/feature';
-import OrderStatuses from '../../constants/order_status';
-import { PAYMENT_METHOD_TYPE } from '../../constants/paymentMethods';
-import { TransactionKind } from '../../constants/transaction-kind';
-import { getFxRate } from '../../lib/currency';
 import logger from '../../lib/logger';
-import { toNegative } from '../../lib/math';
-import { createRefundTransaction, getApplicationFee } from '../../lib/payments';
+import { getApplicationFee } from '../../lib/payments';
 import { reportErrorToSentry, reportMessageToSentry } from '../../lib/sentry';
 import stripe, { convertToStripeAmount } from '../../lib/stripe';
 import models from '../../models';
@@ -257,331 +250,11 @@ export const setupCreditCard = async (
   return paymentMethod;
 };
 
-// Charge has a new Stripe dispute
-async function createDispute(event) {
-  const stripeChargeId = event.data.object.charge;
-  const chargeTransaction = await models.Transaction.findOne({
-    where: { data: { charge: { id: stripeChargeId } } },
-    include: [
-      {
-        model: models.User,
-        required: true,
-        as: 'createdByUser',
-      },
-    ],
-  });
-
-  if (!chargeTransaction) {
-    return;
-  }
-
-  const user = chargeTransaction.createdByUser;
-
-  const transactions = await models.Transaction.findAll({
-    where: {
-      TransactionGroup: chargeTransaction.TransactionGroup,
-    },
-    include: {
-      model: models.Order,
-      required: true,
-      include: [models.Subscription],
-    },
-  });
-
-  // Block User from creating any new Orders
-  await user.limitFeature(FEATURE.ORDER);
-
-  await Promise.all(
-    transactions.map(async transaction => {
-      await transaction.update({ isDisputed: true });
-      const order = transaction.Order;
-      if (order.status !== OrderStatuses.DISPUTED) {
-        await order.update({ status: OrderStatuses.DISPUTED });
-
-        if (order.SubscriptionId) {
-          await order.Subscription.deactivate();
-        }
-      }
-    }),
-  );
-}
-
-// Charge dispute has been closed on Stripe (with status of: won/lost/closed)
-async function closeDispute(event) {
-  const stripeChargeId = event.data.object.charge;
-  const chargeTransaction = await models.Transaction.findOne({
-    where: { data: { charge: { id: stripeChargeId } }, isDisputed: true },
-    include: [
-      {
-        model: models.Order,
-        required: true,
-        include: [models.Subscription],
-      },
-      {
-        model: models.User,
-        required: true,
-        as: 'createdByUser',
-      },
-    ],
-  });
-
-  if (!chargeTransaction) {
-    return;
-  }
-
-  const user = chargeTransaction.createdByUser;
-
-  const transactions = await models.Transaction.findAll({
-    where: {
-      TransactionGroup: chargeTransaction.TransactionGroup,
-    },
-    include: {
-      model: models.Order,
-      required: true,
-      include: [models.Subscription],
-    },
-  });
-
-  if (transactions.length > 0) {
-    const order = chargeTransaction.Order;
-
-    const disputeStatus = event.data.object.status;
-    const disputeTransaction = event.data.object.balance_transactions.find(
-      tx => tx.type === 'adjustment' && tx.reporting_category === 'dispute',
-    );
-
-    // A lost dispute means it was decided as fraudulent
-    if (disputeStatus === 'lost') {
-      if (order.status === OrderStatuses.DISPUTED) {
-        if (order.SubscriptionId) {
-          await order.update({ status: OrderStatuses.CANCELLED });
-        } else {
-          await order.update({ status: OrderStatuses.REFUNDED });
-        }
-      }
-
-      const paymentMethod = await order.getPaymentMethod();
-      if (paymentMethod?.type === PAYMENT_METHOD_TYPE.CREDITCARD) {
-        await models.SuspendedAsset.suspendCreditCard(paymentMethod, `CreditCard was disputed in order ${order.id}`);
-      }
-
-      // Create refund transaction for the fraudulent charge
-      const transactionGroup = uuid();
-      await createRefundTransaction(
-        chargeTransaction,
-        0,
-        {
-          ...chargeTransaction.data,
-          dispute: event,
-          refundTransactionId: chargeTransaction.id,
-        },
-        null,
-        transactionGroup,
-      );
-
-      // Create transaction for dispute fee debiting the fiscal host
-      const feeDetails = disputeTransaction.fee_details.find(feeDetails => feeDetails.description === 'Dispute fee');
-      const currency = feeDetails.currency.toUpperCase();
-      const amount = feeDetails.amount;
-      const fiscalHost = await models.Collective.findByPk(chargeTransaction.HostCollectiveId);
-      const hostCurrencyFxRate = await getFxRate(currency, fiscalHost.currency);
-      const hostCurrencyAmount = Math.round(toNegative(amount) * hostCurrencyFxRate);
-
-      await models.Transaction.create({
-        type: 'DEBIT',
-        HostCollectiveId: fiscalHost.id,
-        CollectiveId: fiscalHost.id,
-        FromCollectiveId: fiscalHost.id,
-        OrderId: chargeTransaction.OrderId,
-        TransactionGroup: transactionGroup,
-        amount: toNegative(amount),
-        netAmountInCollectiveCurrency: toNegative(amount),
-        amountInHostCurrency: hostCurrencyAmount,
-        currency: currency,
-        hostCurrency: fiscalHost.currency,
-        description: 'Stripe Transaction Dispute Fee',
-        paymentProcessorFeeInHostCurrency: 0,
-        hostFeeInHostCurrency: 0,
-        platformFeeInHostCurrency: 0,
-        hostCurrencyFxRate,
-        kind: TransactionKind.PAYMENT_PROCESSOR_DISPUTE_FEE,
-        data: event.data,
-      });
-
-      // A won dispute means it was decided as not fraudulent
-    } else if (disputeStatus === 'won') {
-      if (order.status === OrderStatuses.DISPUTED) {
-        if (order.SubscriptionId) {
-          await order.update({ status: OrderStatuses.ACTIVE });
-          await order.Subscription.activate();
-        } else {
-          await order.update({ status: OrderStatuses.PAID });
-        }
-      }
-
-      const userHasDisputedOrders = await user.hasDisputedOrders();
-      if (!userHasDisputedOrders) {
-        await user.unlimitFeature(FEATURE.ORDER);
-      }
-
-      await Promise.all(
-        transactions.map(async transaction => {
-          await transaction.update({ isDisputed: false });
-        }),
-      );
-    }
-  }
-}
-
-// Charge on Stripe had a fraud review opened
-async function openReview(event) {
-  const stripePaymentIntentId = event.data.object.payment_intent;
-  const paymentIntentTransaction = await models.Transaction.findOne({
-    // eslint-disable-next-line camelcase
-    where: { data: { charge: { payment_intent: stripePaymentIntentId } } },
-    include: [
-      {
-        model: models.Order,
-        required: true,
-        include: [models.Subscription],
-      },
-    ],
-  });
-
-  if (!paymentIntentTransaction) {
-    return;
-  }
-
-  const transactions = await models.Transaction.findAll({
-    where: {
-      TransactionGroup: paymentIntentTransaction.TransactionGroup,
-    },
-    include: {
-      model: models.Order,
-      required: true,
-      include: [models.Subscription],
-    },
-  });
-
-  await Promise.all(
-    transactions.map(async transaction => {
-      await transaction.update({ isInReview: true });
-      const order = transaction.Order;
-      if (order.status !== OrderStatuses.IN_REVIEW) {
-        await order.update({ status: OrderStatuses.IN_REVIEW });
-
-        if (order.SubscriptionId) {
-          await order.Subscription.deactivate();
-        }
-      }
-    }),
-  );
-}
-
-// Charge on Stripe had a fraud review closed (either approved/refunded)
-async function closeReview(event) {
-  const stripePaymentIntentId = event.data.object.payment_intent;
-  const closedReason = event.data.object.closed_reason;
-
-  const paymentIntentTransaction = await models.Transaction.findOne({
-    // eslint-disable-next-line camelcase
-    where: { data: { charge: { payment_intent: stripePaymentIntentId } } },
-    include: [
-      {
-        model: models.Order,
-        required: true,
-        include: [models.Subscription],
-      },
-      {
-        model: models.User,
-        required: true,
-        as: 'createdByUser',
-      },
-    ],
-  });
-
-  if (!paymentIntentTransaction) {
-    return;
-  }
-
-  const user = paymentIntentTransaction.createdByUser;
-
-  const transactions = await models.Transaction.findAll({
-    where: {
-      TransactionGroup: paymentIntentTransaction.TransactionGroup,
-    },
-    include: {
-      model: models.Order,
-      required: true,
-      include: [models.Subscription],
-    },
-  });
-
-  if (transactions.length > 0) {
-    const order = paymentIntentTransaction.Order;
-
-    // closedReasons: approved, refunded, refunded_as_fraud, disputed, redacted
-    if (closedReason === 'refunded_as_fraud' || closedReason === 'refunded') {
-      if (order.status === OrderStatuses.IN_REVIEW) {
-        if (order.SubscriptionId) {
-          await order.update({ status: OrderStatuses.CANCELLED });
-        } else {
-          await order.update({ status: OrderStatuses.REFUNDED });
-        }
-      }
-
-      // Create refund transaction for the fraudulent charge
-      const transactionGroup = uuid();
-      await createRefundTransaction(
-        paymentIntentTransaction,
-        0,
-        {
-          ...paymentIntentTransaction.data,
-          review: event,
-          refundTransactionId: paymentIntentTransaction.id,
-        },
-        null,
-        transactionGroup,
-      );
-
-      // charge review was determined to be fraudulent
-      if (closedReason === 'refunded_as_fraud') {
-        await user.limitFeature(FEATURE.ORDER);
-      } else if (closedReason === 'refunded') {
-        await Promise.all(
-          transactions.map(async transaction => {
-            await transaction.update({ isInReview: false });
-          }),
-        );
-      }
-    } else {
-      if (order.status === OrderStatuses.IN_REVIEW) {
-        if (order.SubscriptionId) {
-          await order.update({ status: OrderStatuses.ACTIVE });
-          await order.Subscription.activate();
-        } else {
-          await order.update({ status: OrderStatuses.PAID });
-        }
-      }
-
-      await Promise.all(
-        transactions.map(async transaction => {
-          await transaction.update({ isInReview: false });
-        }),
-      );
-    }
-  }
-}
-
 export default {
   features: {
     recurring: true,
     waitToCharge: false,
   },
-  closeDispute,
-  createDispute,
-  openReview,
-  closeReview,
 
   processOrder: async order => {
     const hostStripeAccount = await order.collective.getHostStripeAccount();
